requirements.txt
flask==2.0.1
requests==2.28.1
google-generativeai==0.3.2
python-dotenv==0.21.0